requirements.txt
certifi==2020.11.8
chardet==3.0.4
cPython==0.0.6
cycler==0.10.0
Cython==0.29.21
idna==2.10
imageio==2.9.0
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.3
opencv-python==4.4.0.44
Pillow==8.0.1
pymongo==3.11.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.52.0
urllib3==1.26.2
